restructure splitIntoTwoPoints logic
